Move test paths to subpackage
