+++ src/main/java/frc/robot/RobotContainer.java
@@ -25,12 +25,12 @@ import frc.robot.commands.BalancerCommand;
 import frc.robot.commands.sequences.BlueBalanceAuton;
 import frc.robot.commands.sequences.BlueBottomNonBalanceAuton;
 import frc.robot.commands.sequences.BlueTopNonBalanceAuton;
-import frc.robot.commands.sequences.BoxAutonSequence;
-import frc.robot.commands.sequences.GRTAutonSequence;
 import frc.robot.commands.sequences.RedBalanceAuton;
 import frc.robot.commands.sequences.RedBottomNonBalanceAuton;
 import frc.robot.commands.sequences.RedTopNonBalanceAuton;
-import frc.robot.commands.sequences.RotatingSCurveAutonSequence;
+import frc.robot.commands.sequences.test.BoxAutonSequence;
+import frc.robot.commands.sequences.test.GRTAutonSequence;
+import frc.robot.commands.sequences.test.RotatingSCurveAutonSequence;
 import frc.robot.commands.swerve.FollowPathCommand;
 import frc.robot.controllers.BaseDriveController;
 import frc.robot.controllers.DualJoystickDriveController;
