+++ src/main/java/frc/robot/commands/sequences/test/RotatingSCurveAutonSequence.java
@@ -1,4 +1,4 @@
-package frc.robot.commands.sequences;
+package frc.robot.commands.sequences.test;
 
 import java.util.List;
